Adding search and display workflow mockups.
query_builder_for_search_workflow - This workflow is responsible
for building up searchs which can be stored on a per user basis.  The idea is that a user can select an element to query on, and build up the query until they have the set of parameters they want to filter the query down on.  The plus sign on the left is for adding another filter on the fly.

display_counts_workflow - This is an alternate view of the display multiple issues. It is more powerful because it gives you a sense of scope to look at. As it is only displaying counts, it has the advantage of being a bit more efficient than repeatly displaying pages and pages of queries.  Each of the numbers links and can display the lower level ''display_multiple_issues_workflow''

display_multiple_issues_workflow - This workflow is the common page for showing mutliple issues.  We will likely want to count cap the display for this page somehow for performance reasons.  All of the columns are sortable.  The issues link to the ''display_single_bug_workflow''.

Tested: Nothing to be tested here.
Reviewed: Will be at our meeting later today at 5pm.