Fixing tests to run via manage
The tests were not callable via django test due to the
missing __init__.py file.  I found a bug in one of the tests
and correct it.

Also, included in this is the addition of a sidebar for search
and a test to call it via the sidebar.  A new test for search is
still needed.