Fixing edit issue template
Edit issue dropped the 'description' from the submission. The UpdateView
form requires all fields listed to be submitted as part of the update
request.  When we switched from "{{ form }}" to the other edit page
implementations, this field was turned into a text on the page, and
removed it from the form being submitted.