Updating IssueComment model and cleanup
IssueComment - Changed name to be more specific (to avoid collision with
other teams).  Also corrected a couple of bugs in the model. The pk for
the issue is different than the pk for the IssueComment.  Secondly, when
rendering the model via unicode, it expected a string type not a int,
which is what self.pk is.  I fixed it by casting to int.  Lastly, I
exposed this model via the admin site so we could directly add new
comments for specific issues.

Lastly, Pep8/pyflakes cleanup.  Discovered search tests were broken with
the new js calendar plugin with change: 9ad2fdd.  I notified team of the
issue.